Move the outbox test and add a deserialization test
