+++ src/Dafda.Tests/Producing/TestProducer.cs
@@ -311,67 +311,5 @@ namespace Dafda.Tests.Producing
 
             AssertJson.Equal(expected, spy.Value);
         }
-
-        [Fact]
-        public async Task Creates_activity_when_producing_outbox_message()
-        {
-            var topic = "foo";
-            var type = "bar";
-            DafdaActivitySource.Propagator = new CompositeTextMapPropagator(
-                new TextMapPropagator[]
-                {
-                    new TraceContextPropagator(),
-                    new BaggagePropagator()
-                });
-            
-            // Arrange
-            var activitySource = new ActivitySource("ActivitySourceName");
-            using var activity = activitySource.StartActivity("MethodType:/Path");
-            
-            var spy = new KafkaProducerSpy();
-            var outboxProducer = new OutboxProducer(spy);
-            var guid = Guid.NewGuid();
-            
-            var payload = $@"{{
-                                ""messageId"":""{guid.ToString()}"",
-                                ""type"":""{type}"",
-                                ""causationId"":""1"",
-                                ""correlationId"":""1"",                           
-                                ""data"":{{
-                                    ""id"":""dummyId""
-                                    }}
-                                }}";
-
-
-            var outboxEntry = new OutboxEntry(
-                messageId: guid,
-                topic: topic,
-                key: type,
-                payload: payload,
-                occurredUtc: DateTime.UtcNow
-            );
-
-            var activities = new List<Activity>();
-
-            // Act
-            using var activityListener = new ActivityListener
-            {
-                ShouldListenTo = s => s.Name == "Dafda",
-                Sample = (ref ActivityCreationOptions<ActivityContext> _) => ActivitySamplingResult.AllDataAndRecorded,
-                ActivityStarted = activity => activities.Add(activity),
-                ActivityStopped = activity => activities.Add(activity),
-            };
-            ActivitySource.AddActivityListener(activityListener);
-
-            await outboxProducer.Produce(outboxEntry);
-
-            // Assert
-            Assert.Equal("foo", spy.Topic);
-            Assert.Equal("bar", spy.Key);
-            var jsonObject = JObject.Parse(spy.Value);
-            Assert.True(jsonObject["traceparent"] != null, "The JSON does not contain the traceparent element.");
-            Assert.Contains(activities, a => a.DisplayName == $"Dafda.Outbox.{topic}.{type}.{OpenTelemetryMessagingOperation.Producer.Publish}");
-        }
-        
     }
 }
